fix hydration not happening on client-side page transition
